requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
antiorm==1.2.1
attrs==23.1.0
black==23.12.1
cffi==1.16.0
click
connectorx==0.3.2
db==0.1.1
db-sqlite3==0.0.1
discord.py==2.3.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
packaging==23.2
pathspec==0.12.1
peewee==3.17.0
platformdirs==4.1.0
polars==0.20.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyarrow==14.0.2
pycparser==2.21
PyMySQL==1.1.0
PyNaCl==1.5.0
python-dotenv
SQLAlchemy==2.0.24
typing_extensions==4.9.0
yarl==1.9.4
